rename imports to match new file names
