+++ packages/property-tree/src/components/navigation/BuildingList.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import { Property } from '@/services/types'
-import { Skeleton } from '@/components/ui/skeleton'
-import { SidebarMenu } from '@/components/ui/sidebar'
+import { Skeleton } from '@/components/ui/Skeleton'
+import { SidebarMenu } from '@/components/ui/Sidebar'
 import { BuildingNavigation } from './Building'
 import { useQuery } from '@tanstack/react-query'
 import { buildingService } from '@/services/api'
