+++ packages/property-tree/src/components/navigation/ResidenceList.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import { Building } from '@/services/types'
-import { Skeleton } from '@/components/ui/skeleton'
-import { SidebarMenu } from '@/components/ui/sidebar'
+import { Skeleton } from '@/components/ui/Skeleton'
+import { SidebarMenu } from '@/components/ui/Sidebar'
 import { ResidenceNavigation } from './Residence'
 import { useQuery } from '@tanstack/react-query'
 import { residenceService } from '@/services/api'
