+++ packages/property-tree/src/components/navigation/StaircaseList.tsx
@@ -1,6 +1,6 @@
 import { Building } from '@/services/types'
-import { Skeleton } from '@/components/ui/skeleton'
-import { SidebarMenu } from '@/components/ui/sidebar'
+import { Skeleton } from '@/components/ui/Skeleton'
+import { SidebarMenu } from '@/components/ui/Sidebar'
 import { StaircaseNavigation } from './Staircase'
 import { useQuery } from '@tanstack/react-query'
 import { GET } from '@/services/api/baseApi'
