+++ packages/property-tree/src/components/ui/InfoItem.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { LucideIcon } from 'lucide-react'
 
 interface InfoItemProps {
