+++ packages/property-tree/src/components/ui/Sidebar.tsx
@@ -5,19 +5,19 @@ import { Slot } from '@radix-ui/react-slot'
 import { VariantProps, cva } from 'class-variance-authority'
 import { PanelLeft } from 'lucide-react'
 
-import { useIsMobile } from '@/components/hooks/use-mobile'
+import { useIsMobile } from '@/components/hooks/useMobile'
 import { cn } from '@/components/lib/utils/cn'
 import { Button } from '@/components/ui/Button'
-import { Input } from '@/components/ui/input'
-import { Separator } from '@/components/ui/separator'
-import { Sheet, SheetContent } from '@/components/ui/sheet'
-import { Skeleton } from '@/components/ui/skeleton'
+import { Input } from '@/components/ui/Input'
+import { Separator } from '@/components/ui/Separator'
+import { Sheet, SheetContent } from '@/components/ui/Sheet'
+import { Skeleton } from '@/components/ui/Skeleton'
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from '@/components/ui/tooltip'
+} from '@/components/ui/Tooltip'
 
 const SIDEBAR_COOKIE_NAME = 'sidebar:state'
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
